# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Parlor API
  description: API for Parlor salon management and appointment booking system
  version: 1.0.0
  contact:
    name: Parlor Harpia API Docs
servers:
  - url: http://localhost:8080
    description: Local development server
...
paths:
...
  /auth/refresh:
    post:
      tags:
        - Auth
      summary: Refresh access token
      description: Uses the refresh token cookie to issue a new access token
      responses:
        '200':
          description: Token refreshed successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '401':
          description: Refresh token not found or invalid
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    AuthResponse:
      type: object
      properties:
        user:
          $ref: '#/components/schemas/UserResponse'
      description: Authentication response. Tokens are set as HTTP-only cookies, not returned in body.
...
    UserResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        email:
          type: string
          format: email
        name:
          type: string
        phone:
          type: string
        role:
          type: string
          enum: [admin, owner, employee, customer]
        salon_id:
          type: string
          format: uuid
          nullable: true
        onboarding_completed:
          type: boolean
          description: Whether the user has completed onboarding
        current_onboarding_step:
          type: integer
          nullable: true
          description: The user's current onboarding step ID
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
...